# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the API documentation for the Game Service including Player, ClawMachine, GachaMachine, and WhackAMole endpoints.",
        "title": "Game Service API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/clawMachine/addTouchedItemRecord": {
            "post": {
                "description": "Record an item that was touched during a claw game",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ClawMachine"
                ],
                "summary": "Add touched item record",
                "parameters": [
                    {
                        "description": "Add touched item record request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.AddTouchedItemRecordRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Touched item record added successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/clawMachine/adjustPlayerCoin": {
            "post": {
                "description": "Add or subtract coins from a player's balance",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ClawMachine"
                ],
                "summary": "Adjust player coin balance",
                "parameters": [
                    {
                        "description": "Adjust coin request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.AdjustPlayerCoinRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Player coin adjusted successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/clawMachine/adjustPlayerDiamond": {
            "post": {
                "description": "Add or subtract diamonds from a player's balance",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ClawMachine"
                ],
                "summary": "Adjust player diamond balance",
                "parameters": [
                    {
                        "description": "Adjust diamond request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.AdjustPlayerDiamondRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Player diamond adjusted successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/clawMachine/createClawItems": {
            "post": {
                "description": "Create multiple claw items with their properties",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ClawMachine"
                ],
                "summary": "Create claw items",
                "parameters": [
                    {
                        "description": "Claw items creation request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateClawItemsRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Claw items created successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/clawMachine/createClawMachine": {
            "post": {
                "description": "Create a new claw machine with the provided details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ClawMachine"
                ],
                "summary": "Create a new claw machine",
                "parameters": [
                    {
                        "description": "Claw machine creation request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateClawMachineRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Claw machine created successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/clawMachine/createClawPlayer": {
            "post": {
                "description": "Create a new claw player with initial coin and diamond balance",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ClawMachine"
                ],
                "summary": "Create a new claw player",
                "parameters": [
                    {
                        "description": "Claw player creation request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateClawPlayerRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Claw player created successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/clawMachine/deleteClawItems": {
            "post": {
                "description": "Delete multiple claw items by their IDs",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ClawMachine"
                ],
                "summary": "Delete claw items",
                "parameters": [
                    {
                        "description": "Delete claw items request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.DeleteClawItemsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Claw items deleted successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/clawMachine/deleteClawMachine": {
            "post": {
                "description": "Delete a claw machine by machine ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ClawMachine"
                ],
                "summary": "Delete a claw machine",
                "parameters": [
                    {
                        "description": "Delete machine request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.DeleteClawMachineRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Claw machine deleted successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/clawMachine/deleteClawPlayer": {
            "post": {
                "description": "Delete a claw player by player ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ClawMachine"
                ],
                "summary": "Delete a claw player",
                "parameters": [
                    {
                        "description": "Delete player request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.DeleteClawPlayerRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Claw player deleted successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/clawMachine/getClawMachineInfo/:machineID": {
            "get": {
                "description": "Get claw machine information by machine ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ClawMachine"
                ],
                "summary": "Get claw machine information",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Machine ID",
                        "name": "machineID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Claw machine info retrieved successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/clawMachine/getClawPlayerInfo/:playerID": {
            "get": {
                "description": "Get claw player information by player ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ClawMachine"
                ],
                "summary": "Get claw player information",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Player ID",
                        "name": "playerID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Claw player info retrieved successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/clawMachine/getPlayerInventory/{playerID}": {
            "get": {
                "description": "Get inventory for a claw machine player",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ClawMachine"
                ],
                "summary": "Get claw machine player inventory",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Player ID",
                        "name": "playerID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Player inventory retrieved successfully",
                        "schema": {
                            "$ref": "#/definitions/dto.GetClawPlayerInventoryResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid player ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/clawMachine/startClawGame": {
            "post": {
                "description": "Start a new claw game session for a player",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ClawMachine"
                ],
                "summary": "Start a claw game",
                "parameters": [
                    {
                        "description": "Start game request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.StartClawGameRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Claw game started successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/clawMachine/updateClawMachineItems": {
            "post": {
                "description": "Update items in a claw machine",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ClawMachine"
                ],
                "summary": "Update claw machine items",
                "parameters": [
                    {
                        "description": "Update machine items request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateClawMachineItemsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Claw machine items updated successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "dto.AddTouchedItemRecordRequest": {
            "type": "object",
            "required": [
                "catched",
                "gameID",
                "itemID"
            ],
            "properties": {
                "catched": {
                    "type": "boolean"
                },
                "gameID": {
                    "type": "integer"
                },
                "itemID": {
                    "type": "integer"
                }
            }
        },
...
        "dto.AdjustPlayerCoinRequest": {
            "type": "object",
            "required": [
                "amount",
                "playerID",
                "type"
            ],
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "playerID": {
                    "type": "integer"
                },
                "type": {
                    "type": "string",
                    "enum": [
                        "plus",
                        "minus"
                    ]
                }
            }
        },
        "dto.AdjustPlayerDiamondRequest": {
            "type": "object",
            "required": [
                "amount",
                "playerID",
                "type"
            ],
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "playerID": {
                    "type": "integer"
                },
                "type": {
                    "type": "string",
                    "enum": [
                        "plus",
                        "minus"
                    ]
                }
            }
        },
...
        "dto.CreateClawItemsRequest": {
            "type": "object",
            "required": [
                "clawItems"
            ],
            "properties": {
                "clawItems": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.CreateClawItemRequest"
                    }
                }
            }
        },
...
        "dto.CreateClawMachineRequest": {
            "type": "object",
            "required": [
                "maxItem",
                "name",
                "price"
            ],
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.CreateClawMachineItemRequest"
                    }
                },
                "maxItem": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "integer"
                }
            }
        },
...
        "dto.DeleteClawItemsRequest": {
            "type": "object",
            "required": [
                "itemIDs"
            ],
            "properties": {
                "itemIDs": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
...
    }
}